# Conflicts:
#	_maps/map_files/generic/Lavaland.dmm
#	_maps/map_files/generic/centcomm.dmm
#	_maps/map_files/templates/medium_shuttle2.dmm
#	code/controllers/subsystem/ticker.dm
#	code/datums/mind.dm
#	code/game/machinery/OpTable.dm
#	code/game/machinery/Sleeper.dm
#	code/game/machinery/cell_charger.dm
#	code/game/machinery/computer/card.dm
#	code/game/machinery/computer/cloning.dm
#	code/game/machinery/computer/prisoner.dm
#	code/game/machinery/constructable_frame.dm
#	code/game/machinery/doors/airlock_control.dm
#	code/game/machinery/embedded_controller/airlock_controllers.dm
#	code/game/machinery/embedded_controller/embedded_controller_base.dm
#	code/game/machinery/firealarm.dm
#	code/game/machinery/holosign.dm
#	code/game/machinery/lightswitch.dm
#	code/game/machinery/mass_driver.dm
#	code/game/machinery/pipe/construction.dm
#	code/game/machinery/recharger.dm
#	code/game/machinery/suit_storage_unit.dm
#	code/game/machinery/syndicatebeacon.dm
#	code/game/objects/items/weapons/storage/secure.dm
#	code/game/objects/items/weapons/storage/storage.dm
#	code/game/objects/objs.dm
#	code/game/objects/structures/barsign.dm
#	code/game/objects/structures/crates_lockers/closets/secure/secure_closets.dm
#	code/game/objects/structures/displaycase.dm
#	code/game/objects/structures/door_assembly.dm
#	code/game/objects/structures/extinguisher.dm
#	code/game/objects/structures/mirror.dm
#	code/game/objects/structures/morgue.dm
#	code/game/objects/structures/signs.dm
#	code/game/objects/structures/tank_dispenser.dm
#	code/game/objects/structures/windoor_assembly.dm
#	code/game/turfs/simulated/walls_reinforced.dm
#	code/modules/atmospherics/machinery/airalarm.dm
#	code/modules/atmospherics/machinery/atmospherics.dm
#	code/modules/atmospherics/machinery/components/unary_devices/cryo.dm
#	code/modules/atmospherics/machinery/components/unary_devices/vent_pump.dm
#	code/modules/atmospherics/machinery/components/unary_devices/vent_scrubber.dm
#	code/modules/atmospherics/machinery/other/meter.dm
#	code/modules/atmospherics/machinery/portable/canister.dm
#	code/modules/atmospherics/machinery/portable/portable_atmospherics.dm
#	code/modules/clothing/spacesuits/hardsuit.dm
#	code/modules/clothing/spacesuits/plasmamen.dm
#	code/modules/mob/living/carbon/human/human.dm
#	code/modules/mob/living/living_defines.dm
#	code/modules/paperwork/folders.dm
#	code/modules/paperwork/paper_bundle.dm
#	code/modules/pda/PDA.dm
#	code/modules/persistence/persistence.dm
#	code/modules/power/cable.dm
#	code/modules/power/cell.dm
#	code/modules/power/lighting.dm
#	code/modules/power/port_gen.dm
#	code/modules/projectiles/ammunition.dm
#	code/modules/projectiles/gun.dm
#	code/modules/reagents/chemistry/machinery/chem_master.dm
#	code/modules/reagents/reagent_containers.dm
#	code/modules/recycling/disposal-construction.dm
#	code/modules/recycling/disposal.dm
#	code/modules/surgery/organs/organ.dm
#	paradise.dme

+++ code/game/turfs/turf.dm
@@ -27,7 +27,7 @@
 	//Properties for both
 	var/temperature = T20C
 
-	var/blocks_air = 0
+	var/blocks_air = FALSE
 
 	var/datum/pathnode/PNode = null //associated PathNode in the A* algorithm
 
@@ -54,6 +54,10 @@
 	proc/deserialize_air(list/data)
 		return
 
+
+	/// How pathing algorithm will check if this turf is passable by itself (not including content checks). By default it's just density check.
+	/// WARNING: Currently to use a density shortcircuiting this does not support dense turfs with special allow through function
+	var/pathing_pass_method = TURF_PATHING_PASS_DENSITY
 
 /turf/Initialize(mapload)
 	SHOULD_CALL_PARENT(FALSE)
@@ -108,6 +112,7 @@
 		for(var/A in B.contents)
 			qdel(A)
 		return
+	REMOVE_FROM_SMOOTH_QUEUE(src)
 	// Adds the adjacent turfs to the current atmos processing
 	for(var/turf/simulated/T in atmos_adjacent_turfs)
 		SSair.add_to_active(T)
@@ -219,7 +224,7 @@
 // Removes all signs of lattice on the pos of the turf -Donkieyo
 /turf/proc/RemoveLattice()
 	var/obj/structure/lattice/L = locate(/obj/structure/lattice, src)
-	if(L)
+	if(L && !(L.resistance_flags & INDESTRUCTIBLE))
 		qdel(L)
 
 /turf/proc/dismantle_wall(devastated = FALSE, explode = FALSE)
@@ -229,7 +234,7 @@
 	return ChangeTurf(path, defer_change, keep_icon, ignore_air)
 
 //Creates a new turf
-/turf/proc/ChangeTurf(path, defer_change = FALSE, keep_icon = TRUE, ignore_air = FALSE)
+/turf/proc/ChangeTurf(path, defer_change = FALSE, keep_icon = TRUE, ignore_air = FALSE, copy_existing_baseturf = TRUE)
 	if(!path)
 		return
 	if(!GLOB.use_preloader && path == type) // Don't no-op if the map loader requires it to be reconstructed
@@ -250,7 +255,8 @@
 	changing_turf = TRUE
 	qdel(src)	//Just get the side effects and call Destroy
 	var/turf/W = new path(src)
-	W.baseturf = old_baseturf
+	if(copy_existing_baseturf)
+		W.baseturf = old_baseturf
 
 	if(!defer_change)
 		W.AfterChange(ignore_air)
@@ -290,7 +296,7 @@
 	levelupdate()
 	CalculateAdjacentTurfs()
 
-	if(SSair && !ignore_air)
+	if(!ignore_air)
 		SSair.add_to_active(src)
 
 	//update firedoor adjacency
@@ -324,10 +330,10 @@
 		var/turf_count = 0
 
 		for(var/turf/T in atmos_adjacent_turfs)
-			if(istype(T, /turf/space))//Counted as no air
+			if(isspaceturf(T))//Counted as no air
 				turf_count++//Considered a valid turf for air calcs
 				continue
-			else if(istype(T, /turf/simulated/floor))
+			else if(isfloorturf(T))
 				var/turf/simulated/S = T
 				if(S.air)//Add the air's contents to the holders
 					aoxy += S.air.oxygen
@@ -360,9 +366,9 @@
 	for(var/mob/living/M in src)
 		if(M == U)
 			continue//Will not harm U. Since null != M, can be excluded to kill everyone.
-		INVOKE_ASYNC(M, /mob/.proc/gib)
+		INVOKE_ASYNC(M, TYPE_PROC_REF(/mob, gib))
 	for(var/obj/mecha/M in src)//Mecha are not gibbed but are damaged.
-		INVOKE_ASYNC(M, /obj/mecha/.proc/take_damage, 100, "brute")
+		INVOKE_ASYNC(M, TYPE_PROC_REF(/obj/mecha, take_damage), 100, "brute")
 
 /turf/proc/Bless()
 	flags |= NOJAUNT
@@ -510,10 +516,7 @@
 /turf/proc/acid_melt()
 	return
 
-/turf/handle_fall(mob/faller, forced)
-	faller.lying = pick(90, 270)
-	if(!forced)
-		return
+/turf/handle_fall()
 	if(has_gravity(src))
 		playsound(src, "bodyfall", 50, TRUE)
 
@@ -601,3 +604,24 @@
 
 /turf/AllowDrop()
 	return TRUE
+
+/**
+ * Returns adjacent turfs to this turf that are reachable, in all cardinal directions
+ *
+ * Arguments:
+ * * caller: The movable, if one exists, being used for mobility checks to see what tiles it can reach
+ * * ID: An ID card that decides if we can gain access to doors that would otherwise block a turf
+ * * simulated_only: Do we only worry about turfs with simulated atmos, most notably things that aren't space?
+ * * no_id: When true, doors with public access will count as impassible
+*/
+/turf/proc/reachableAdjacentTurfs(caller, ID, simulated_only, no_id = FALSE)
+	var/static/space_type_cache = typecacheof(/turf/space)
+	. = list()
+
+	for(var/iter_dir in GLOB.cardinal)
+		var/turf/turf_to_check = get_step(src, iter_dir)
+		if(!turf_to_check || (simulated_only && space_type_cache[turf_to_check.type]))
+			continue
+		if(turf_to_check.density || LinkBlockedWithAccess(turf_to_check, caller, ID, no_id = no_id))
+			continue
+		. += turf_to_check
